backend/requirements.txt
neo4j==5.15.0
python-dotenv==1.0.0
pydantic>=2.9.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
nltk>=3.8.1
pytest>=7.4.0
httpx>=0.24.0
google-generativeai>=0.8.0
python-Levenshtein>=0.21.0
Fuzzy>=1.2.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
stripe>=14.0.0
supabase>=2.0.0
PyJWT>=2.8.0
fsrs>=4.0.0

